(refactor) modularize code
