+++ js/detail/domElements.js
@@ -1,0 +1,2 @@
+export const reviewForm = document.querySelector(".review__form");
+export const reviewContainer = document.querySelector(".review__container");
